# Make Poly1305 faster by leveraging @addWithOverflow/@subWithOverflow